+++ src/game/platform_xbox/MPLiveEngine.h
@@ -371,8 +371,8 @@ public:
     void ClearKeys();
     void DoWork();
     void FreeSlot(bool isPrivate);
-    void LockPort(unsigned int port); // ?LockPort@MPLiveEngine@@QAEXK@Z (game2.o 0x4EBE10)
-    unsigned int GetLockedPort();  // ?GetLockedPort@MPLiveEngine@@QAEKXZ (sv.o 0x51E190)
+    void LockPort(unsigned long port); // ?LockPort@MPLiveEngine@@QAEXK@Z (game2.o 0x4EBE10)
+    unsigned long GetLockedPort();  // ?GetLockedPort@MPLiveEngine@@QAEKXZ (sv.o 0x51E190)
     static MPLiveEngine* GetHandle();
     LivePlayer* GetLocalPlayer();  // shell.o 0x5AD260
     unsigned int GetPortToLock();
